+++ apps/auto-email-sender/roles-assigner/src/tests/helpers/stubs.ts
@@ -1,8 +1,8 @@
+import * as getAllPaginatedFile from 'db/src/supabase/getAllPaginated'
 import { SupabaseTable } from 'db/src/supabase/utilityTypes'
 import { mockAsyncGenerator } from 'shared/src/test/helpers/mockAsyncGeneratorFunction'
 import { supabaseClientMock } from 'shared/src/test/helpers/mocks'
 import { Mock, vi } from 'vitest'
-import * as getAllPaginatedFile from '../../getAllPaginated'
 import * as getRowsBlockFile from '../../getRowsBlock'
 
 type Role = SupabaseTable<'Roles'>
